Versie 16, Bestellingen + BesteldeItems aanmaken gelukt in de database. Nu nog mergen.

+++ DAL/MenuItemDao.cs
@@ -36,7 +36,7 @@ namespace DAL
             {
                 MenuItemId = Convert.ToInt32(row["MenuItemId"]),
                 Voorraad = Convert.ToInt32(row["Voorraad"]),
-                Prijs = (decimal)row["Prijs"],
+                Prijs = (double)row["Prijs"],
                 Naam = (string)row["Naam"],
                 MenuId = Convert.ToInt32(row["MenuId"]),
                 IsAlcoholisch = (bool)row["Alcoholisch"],
